chapter-list: refactored and added new features optimize scroll-list.lua: add format_header_string  function for custom behaviour. add more methods and keybinds meant to control the list overlay.
refactor chapter-list.lua:
add mouse control and options.
display the cursor position and the total number of lists in the header.
when the menu is first opened now the new behavior of list.selected is to select the current chapter.
reset cursor navigation when open the list, also add option to control whether this behavior is enabled(default).